refactor: unified filters in new file
